+++ internal/librarian/internal/rust/codec_test.go
