+++ internal/librarian/release.go
@@ -16,9 +16,10 @@ package librarian
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/googleapis/librarian/internal/config"
-	"github.com/googleapis/librarian/internal/language"
+	"github.com/googleapis/librarian/internal/librarian/internal/rust"
 	"github.com/urfave/cli/v3"
 )
 
@@ -65,12 +66,34 @@ func runRelease(ctx context.Context, cmd *cli.Command) error {
 		return err
 	}
 	if all {
-		cfg, err = language.ReleaseAll(cfg)
+		cfg, err = releaseAll(cfg)
 	} else {
-		cfg, err = language.ReleaseLibrary(cfg, libraryName)
+		cfg, err = releaseLibrary(cfg, libraryName)
 	}
 	if err != nil {
 		return err
 	}
 	return cfg.Write(librarianConfigPath)
+}
+
+func releaseAll(cfg *config.Config) (*config.Config, error) {
+	switch cfg.Language {
+	case "testhelper":
+		return testReleaseAll(cfg)
+	case "rust":
+		return rust.ReleaseAll(cfg)
+	default:
+		return nil, fmt.Errorf("language not supported for release --all: %q", cfg.Language)
+	}
+}
+
+func releaseLibrary(cfg *config.Config, name string) (*config.Config, error) {
+	switch cfg.Language {
+	case "testhelper":
+		return testReleaseLibrary(cfg, name)
+	case "rust":
+		return rust.ReleaseLibrary(cfg, name)
+	default:
+		return nil, fmt.Errorf("language not supported for release --all: %q", cfg.Language)
+	}
 }
